--- app/src/main/java/za/co/lbnkosi/portfolio/ui/features/addresses/AddressFragment.kt
@@ -1,4 +1,0 @@
-package za.co.lbnkosi.portfolio.ui.features.addresses
-
-class AddressFragment {
-}
