--- app/src/main/java/za/co/lbnkosi/portfolio/ui/features/competencies/CompetencyFragment.kt
@@ -1,4 +1,0 @@
-package za.co.lbnkosi.portfolio.ui.features.competencies
-
-class CompetencyFragment {
-}
