--- app/src/main/java/za/co/lbnkosi/portfolio/ui/features/socials/SocialFragment.kt
@@ -1,4 +1,0 @@
-package za.co.lbnkosi.portfolio.ui.features.socials
-
-class SocialFragment {
-}
